^ code update with cli args support ^ move exec to helper ^ improve flickr class + add cache

+++ xgallery/libraries/cli/flickr/download.php
@@ -19,58 +19,59 @@ class XgalleryCliFlickrDownload extends JApplicationCli
 	public function doExecute()
 	{
 		\Joomla\CMS\Factory::$application = $this;
-		$input = \Joomla\CMS\Factory::getApplication()->input->cli;
+		$input                            = \Joomla\CMS\Factory::getApplication()->input->cli;
 
 		$db  = \Joomla\CMS\Factory::getDbo();
 		$pid = $input->get('pid');
 
-		XgalleryHelperLog::getLogger()->info('Download photo : ' . $pid);
-
-		try
+		if ($pid)
 		{
-			$db->transactionStart();
-			$query = ' SELECT ' . $db->quoteName('urls') . ',' . $db->quoteName('owner')
-				. ' FROM ' . $db->quoteName('#__xgallery_flickr_contact_photos')
-				. ' WHERE ' . $db->quoteName('id') . ' = ' . $db->quote($pid)
-				. ' LIMIT 1 FOR UPDATE ';
-			$photo = $db->setQuery($query)->loadObject();
-			$urls  = json_decode($photo->urls);
-			$url   = end($urls->sizes->size);
-
-			$toDir = JPATH_ROOT . '/media/xgallery/' . $photo->owner;
-			\Joomla\Filesystem\Folder::create($toDir);
-			$fileName         = basename($url->source);
-			$saveTo           = $toDir . '/' . $fileName;
-			$originalFileSize = XgalleryHelperFile::downloadFile($url->source, $saveTo);
-
-			if ($originalFileSize === false || $originalFileSize != filesize($saveTo))
+			try
 			{
-				\Joomla\Filesystem\File::delete($saveTo);
+				$db->transactionStart();
+				$query = ' SELECT ' . $db->quoteName('urls') . ',' . $db->quoteName('owner')
+					. ' FROM ' . $db->quoteName('#__xgallery_flickr_contact_photos')
+					. ' WHERE ' . $db->quoteName('id') . ' = ' . $db->quote($pid)
+					. ' LIMIT 1 FOR UPDATE ';
+				$photo = $db->setQuery($query)->loadObject();
+				$urls  = json_decode($photo->urls);
+				$url   = end($urls->sizes->size);
 
-				throw new Exception('Download failed: ' . filesize($saveTo) . '/' . $originalFileSize);
+				$toDir = JPATH_ROOT . '/media/xgallery/' . $photo->owner;
+				\Joomla\Filesystem\Folder::create($toDir);
+				$fileName         = basename($url->source);
+				$saveTo           = $toDir . '/' . $fileName;
+				$originalFileSize = XgalleryHelperFile::downloadFile($url->source, $saveTo);
+
+				if ($originalFileSize === false || $originalFileSize != filesize($saveTo))
+				{
+					\Joomla\Filesystem\File::delete($saveTo);
+
+					throw new Exception('Download failed');
+				}
+				else
+				{
+					$query = $db->getQuery(true);
+					// Update this photo status
+					$query->clear()
+						->update($db->quoteName('#__xgallery_flickr_contact_photos'))
+						->set(array(
+							$db->quoteName('state') . ' = 2'
+						))
+						->where($db->quoteName('id') . ' = ' . $db->quote($pid));
+					$db->setQuery($query)->execute();
+
+					$db->transactionCommit();
+
+					XgalleryHelperLog::getLogger()->info('---- Download completed ' . $pid . ' ----');
+				}
+
 			}
-			else
+			catch (Exception $exception)
 			{
-				$query = $db->getQuery(true);
-				// Update this photo status
-				$query->clear()
-					->update($db->quoteName('#__xgallery_flickr_contact_photos'))
-					->set(array(
-						$db->quoteName('state') . ' = 2'
-					))
-					->where($db->quoteName('id') . ' = ' . $db->quote($pid));
-				$db->setQuery($query)->execute();
-
-				$db->transactionCommit();
-
-				XgalleryHelperLog::getLogger()->info('---- Download completed ' . $pid . ' ----', array('saveTo' => $saveTo));
+				XgalleryHelperLog::getLogger()->error($exception->getMessage(), array('query' => (string) $query, 'url' => get_object_vars($urls)));
+				$db->transactionRollback();
 			}
-
-		}
-		catch (Exception $exception)
-		{
-			XgalleryHelperLog::getLogger()->error($exception->getMessage(), array('query' => (string) $query, 'url' => get_object_vars($urls)));
-			$db->transactionRollback();
 		}
 
 		$db->disconnect();
